refactor: 태그가져오는 액션 훅으로 분리
